// workshop-requisition-codeapp/src/generated/models/Wksp_costcentresModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Wksp_costcentresstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Wksp_costcentresstatecode = keyof typeof Wksp_costcentresstatecode;
export const Wksp_costcentresstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Wksp_costcentresstatuscode = keyof typeof Wksp_costcentresstatuscode;

export interface Wksp_costcentresBase {
  importsequencenumber?: number;
  overriddencreatedon?: string;
  statecode: Wksp_costcentresstatecode;
  statuscode?: Wksp_costcentresstatuscode;
  timezoneruleversionnumber?: number;
  "TransactionCurrencyId@odata.bind"?: string;
  utcconversiontimezonecode?: number;
  wksp_committedamount?: number;
  wksp_costcentreid: string;
  wksp_monthlybudget?: number;
  wksp_name?: string;
}

export interface Wksp_costcentres extends Wksp_costcentresBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  exchangerate?: number;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  ownerid: string;
  owneridname: string;
  owneridtype: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  transactioncurrencyidname?: string;
  versionnumber?: number;
  wksp_committedamount_base?: number;
  wksp_monthlybudget_base?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
  transactioncurrencyid?: object;
  _transactioncurrencyid_value?: string;
}